+++ app/api/response/userResponse.ts
@@ -1,0 +1,5 @@
+export interface UserResponse {
+  email: string;
+  password: string;
+  rank: string;
+}
