+++ app/hooks/pageHooks/README.md
@@ -1,0 +1,1 @@
+ページ固有のフックを配置。
