+++ app/repositories/IUserRepository.ts
@@ -1,0 +1,5 @@
+import { UserResponse } from "@/app/api/response/userResponse";
+
+export interface IUserRepository {
+  fetchUserData(): Promise<UserResponse[]>;
+}
